# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Setlists API",
        "contact": {
            "name": "Support Team",
            "email": "support@example.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/song": {
            "post": {
                "tags": [
                    "Songs"
                ],
                "description": "Creates a song.",
                "operationId": "App\\Http\\Controllers\\SongController::createSong",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "A valid and recent token.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "description": "A valid Version 4 Uuid",
                                        "type": "string"
                                    },
                                    "title": {
                                        "description": "The title of the song",
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "id": 10,
                                    "title": "Example Song"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "The song has been succesfully created.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "409": {
                        "description": "Error: id or title already exists."
                    },
                    "500": {
                        "description": "Error: invalid identifier or invalid title."
                    },
                    "400": {
                        "$ref": "#/components/responses/expired"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthorized"
                    }
                }
            }
        },
        "/api/song/{uuid}": {
            "get": {
                "tags": [
                    "Songs"
                ],
                "description": "Returns a song.",
                "operationId": "App\\Http\\Controllers\\SongController::getSong",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "A valid and recent token.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "The uuid of the song.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A song with all its attributes.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Error: invalid identifier."
                    },
                    "404": {
                        "description": "Error: the requested song does not exist."
                    },
                    "400": {
                        "$ref": "#/components/responses/expired"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthorized"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Songs"
                ],
                "description": "Deletes a song.",
                "operationId": "App\\Http\\Controllers\\SongController::deleteSong",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "A valid and recent token.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "The uuid of the song.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The song has been succesfully deleted.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Error: the requested song does not exist."
                    },
                    "500": {
                        "description": "Error: invalid identifier or the song is present in one or more setlists."
                    },
                    "400": {
                        "$ref": "#/components/responses/expired"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthorized"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Songs"
                ],
                "description": "Updates the fields 'title' and 'visibility' from a song.",
                "operationId": "App\\Http\\Controllers\\SongController::updateSong",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "A valid and recent token.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "The uuid of the song.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "description": "The title of the song",
                                        "type": "string"
                                    },
                                    "visibility": {
                                        "description": "Hide or unhide the song.",
                                        "type": "integer"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "title": "Example Song",
                                    "visibility": 0
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "The song has been succesfully updated.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Error: the requested song does not exist."
                    },
                    "500": {
                        "description": "Error: invalid identifier or invalid title."
                    },
                    "400": {
                        "$ref": "#/components/responses/expired"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthorized"
                    }
                }
            }
        },
        "/api/song/{uuid}/force": {
            "delete": {
                "tags": [
                    "Songs"
                ],
                "description": "Deletes a song even if it belongs to one or more setlists.",
                "operationId": "App\\Http\\Controllers\\SongController::forceDeleteSong",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "A valid and recent token.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "The uuid of the song.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "The song has been succesfully deleted.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Error: the requested song does not exist."
                    },
                    "500": {
                        "description": "Error: invalid identifier."
                    },
                    "400": {
                        "$ref": "#/components/responses/expired"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthorized"
                    }
                }
            }
        },
        "/api/songs": {
            "get": {
                "tags": [
                    "Songs"
                ],
                "description": "Returns all stored songs or a range of them, if the optional parameter 'interval' was provided. Filtering by song title is also possible.",
                "operationId": "App\\Http\\Controllers\\SongController::getSongs",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "A valid and recent token.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "interval",
                        "in": "query",
                        "description": "The offset and limit of the requested songs collection, separated with a comma.",
                        "style": "form",
                        "explode": "false",
                        "allowReserved": "true",
                        "schema": {
                            "type": "string"
                        },
                        "examples": [
                            {
                                "summary": "Getting only the first song",
                                "value": "0,1"
                            },
                            {
                                "summary": "A list of songs, starting in the 11th song and with a length of 50 songs.",
                                "value": "10,50"
                            }
                        ]
                    },
                    {
                        "name": "title",
                        "in": "query",
                        "description": "A string to filter the title of the songs in the collection. The comparison is case insensitive.",
                        "style": "form",
                        "explode": "false",
                        "allowReserved": "true",
                        "schema": {
                            "type": "string"
                        },
                        "examples": [
                            {
                                "summary": "Getting only those songs with the string 'example' in their titles.",
                                "value": "example"
                            }
                        ]
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A collection of songs with all their attributes, according to the 'interval' and 'name' parameters, if provided.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "$ref": "#/components/responses/expired"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthorized"
                    }
                }
            }
        },
        "/api/song/stats/{uuid}": {
            "get": {
                "tags": [
                    "Songs"
                ],
                "description": "Returns the setlist to which this song belongs.",
                "operationId": "App\\Http\\Controllers\\SongController::getSongStats",
                "parameters": [
                    {
                        "name": "token",
                        "in": "query",
                        "description": "A valid and recent token.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "The uuid of the song.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A collection of setlists, without acts.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "500": {
                        "description": "Error: invalid identifier."
                    },
                    "404": {
                        "description": "Error: the requested song does not exist."
                    },
                    "400": {
                        "$ref": "#/components/responses/expired"
                    },
                    "401": {
                        "$ref": "#/components/responses/unauthorized"
                    }
                }
            }
        }
    },
    "components": {
        "responses": {
            "expired": {
                "description": "Error: the provided token has expired, or is invalid or is obsolete."
            },
            "unauthorized": {
                "description": "Error: unauthorized request (no token present)"
            },
...
        }
    },
...
}